requirements.txt
accelerate==0.30.1
antialiased-cnns==0.3
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
clean-fid==0.1.35
click==8.1.7
cmake==3.31.1
cupy-cuda12x==13.3.0
diffusers==0.25.1
docker-pycreds==0.4.0
fastrlock==0.8.2
filelock==3.16.1
fsspec==2024.10.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.23.5
idna==3.10
imagecodecs==2024.9.22
importlib_metadata==8.5.0
Jinja2==3.1.4
lit==18.1.8
lpips==0.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.2
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
openslide-bin==4.0.0.6
openslide-python==1.4.1
packaging==24.2
peft==0.10.0
pillow==10.3.0
platformdirs==4.3.6
protobuf==5.29.1
psutil==6.1.0
pycparser==2.22
PySocks==1.7.1
pyvips==2.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scipy==1.14.1
sentry-sdk==2.19.0
setproctitle==1.3.4
six==1.17.0
smmap==5.0.1
soupsieve==2.6
sympy==1.13.1
tifffile==2024.9.20
timm==1.0.12
tokenizers==0.15.2
torch==2.5.1+cu121
torchvision==0.20.1
tqdm==4.66.6
transformers==4.35.2
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
wandb==0.17.9
xformers==0.0.28.post3
zipp==3.21.0
